+++ CHANGELOG.md
@@ -32,6 +32,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 - Admin's /blocking command
 
+### Updated
+
+- Refactor NesUser model & schema
+
 ## [0.1.0] - 2025-06-11 - Vlad
 
 ### Added